+++ Api/Models/DTO/Request/Authorization/ResetPassword/ResetPasswordRequest.cs
@@ -5,7 +5,20 @@ namespace Api.Models.DTO.Request.Authorization.ResetPassword
     public class ResetPasswordRequest
     {
         [Required(ErrorMessage = "Field can't be empty")]
-        [DataType(DataType.EmailAddress, ErrorMessage = "E-mail is not valid")]
-        public string Email { get; set; }
+        public string UserId { get; set; }
+
+        [Required(ErrorMessage = "Field can't be empty")]
+        public string Token { get; set; }
+
+        [Required(ErrorMessage = "Password is required")]
+        [StringLength(255, ErrorMessage = "Must be between 5 and 255 characters", MinimumLength = 5)]
+        [DataType(DataType.Password)]
+        public string Password { get; set; }
+
+        [Required(ErrorMessage = "Confirm Password is required")]
+        [StringLength(255, ErrorMessage = "Must be between 5 and 255 characters", MinimumLength = 5)]
+        [DataType(DataType.Password)]
+        [Compare("Password")]
+        public string PasswordConfirm { get; set; }
     }
 }
