--- Api/Models/DTO/Request/Authorization/ResetPassword/SetPasswordRequest.cs
@@ -1,25 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Api.Models.DTO.Request.Authorization.ResetPassword
-{
-    public class SetPasswordRequest
-    {
-        [Required(ErrorMessage = "Field can't be empty")]
-        [DataType(DataType.EmailAddress, ErrorMessage = "E-mail is not valid")]
-        public string Email { get; set; }
-
-        [Required(ErrorMessage = "Field can't be empty")]
-        public string Code { get; set; }
-
-        [Required(ErrorMessage = "Password is required")]
-        [StringLength(255, ErrorMessage = "Must be between 5 and 255 characters", MinimumLength = 5)]
-        [DataType(DataType.Password)]
-        public string Password { get; set; }
-
-        [Required(ErrorMessage = "Confirm Password is required")]
-        [StringLength(255, ErrorMessage = "Must be between 5 and 255 characters", MinimumLength = 5)]
-        [DataType(DataType.Password)]
-        [Compare("Password")]
-        public string PasswordConfirm { get; set; }
-    }
-}
