+++ Api/appsettings.json
@@ -16,6 +16,11 @@
     "Host": "smtp.gmail.com",
     "Port": 587
   },
-  "Client": "http://localhost:19000",
-  "OpenStreetMap": "https://nominatim.openstreetmap.org"
+  "Authentication": {
+    "Google": {
+      "ClientId": "123",
+      "ClientSecret": "321"
+    }
+  },
+  "TravelSpotApp": "http://localhost:19000"
 }
